move startTransport and helper functions to nodeserver.ts
