+++ src/nodeserver.ts
@@ -1,6 +1,19 @@
+import * as http from "http";
+import * as https from "https";
+import * as net from "net";
+import * as ws from "ws";
 
-import { UserRights } from "./hub";
+import Promise from "ts-promise";
+import Hub, { UserRights } from "./hub";
 import { TlsOptions } from "./tls";
+import TcpConnection from "./transports/tcpconnection";
+import WSConnection from "./transports/wsconnection";
+
+import log from "./log";
+
+const DEFAULT_PORT_WS = 13900;
+const DEFAULT_PORT_WSS = 13901;
+const DEFAULT_PORT_TCP = 13902;
 
 export interface Binding {
     from: string;
@@ -41,4 +54,83 @@ export interface Config {
     storage?: string;
     users?: string | { [username: string]: string };
     rights: UserRights;
+}
+
+// Initialize and start server
+
+let connectionId = 0;
+function startWebSocketServer(hub: Hub, options: WSServerOptions): Promise<void> {
+    return new Promise<void>((resolve, reject) => {
+        options = { ...options }; // clone
+
+        let server: http.Server | https.Server;
+        const useTls = !!(options.key || options.pfx);
+
+        options.port = options.port || (useTls ? DEFAULT_PORT_WS : DEFAULT_PORT_WSS);
+
+        if (useTls) {
+            server = https.createServer(options);
+        } else {
+            server = http.createServer();
+        }
+
+        const wss = new ws.Server({ server: <any>server, path: "/" });
+        wss.on("connection", (conn: ws) => {
+            // tslint:disable-next-line:no-unused-expression
+            new WSConnection(hub, conn, "websocket" + connectionId++);
+        });
+
+        server.listen(options.port, (): void => {
+            log.info("WebSocket Server started on port " + options.port, useTls ? "(TLS)" : "");
+            resolve(undefined);
+        });
+
+        server.on("error", (e: Error): void => {
+            reject(e);
+        });
+    });
+}
+
+function startTcpServer(hub: Hub, options: TcpServerOptions): Promise<void> {
+    return new Promise<void>((resolve, reject) => {
+        options = { ...options }; // clone
+        options.port = options.port || DEFAULT_PORT_TCP;
+
+        const server = net.createServer((socket: net.Socket) => {
+            // tslint:disable-next-line:no-unused-expression
+            new TcpConnection(hub, socket, "tcp" + connectionId++);
+        });
+
+        server.listen(
+            {
+                port: options.port,
+                host: options.host,
+                backlog: options.backlog,
+            },
+            (): void => {
+                log.info("TCP Server started on port " + options.port);
+                resolve(undefined);
+            }
+        );
+
+        server.on("error", (e: Error): void => {
+            reject(e);
+        });
+    });
+}
+
+export function startTransports(hub: Hub, config: Config): Promise<void> {
+    const serverOptions = Array.isArray(config.listen) ? config.listen : [config.listen];
+    return Promise.all(
+        serverOptions.map((options: ListenOptions) => {
+            switch (options.type) {
+                case "websocket":
+                    return startWebSocketServer(hub, <WSServerOptions>options);
+                case "tcp":
+                    return startTcpServer(hub, <TcpServerOptions>options);
+                default:
+                    throw new Error(`unsupported transport '${options!.type}'`);
+            }
+        })
+    ).return();
 }
